fastapi/requirements_fastapi_ai.txt
accelerate==0.23.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
diffusers==0.32.2
fastapi==0.99.1
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pillow==10.4.0
pydantic==1.10.21
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
starlette==0.27.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
tqdm==4.67.1
transformers==4.33.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==11.0.3
info-nce-pytorch==0.1.0
kornia==0.6.8
pygame==2.5.0
